@@ -14,19 +14,20 @@
   "devDependencies": {
     "@eslint/js": "^9.25.1",
     "@fontsource-variable/inter": "^5.2.5",
-    "@sveltejs/adapter-auto": "^3.3.1",
+    "@sveltejs/adapter-auto": "^6.0.0",
     "@sveltejs/adapter-node": "^5.2.12",
     "@sveltejs/adapter-static": "^3.0.8",
-    "@sveltejs/adapter-vercel": "^5.7.0",
-    "@sveltejs/kit": "^2.20.7",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/adapter-vercel": "^5.7.2",
+    "@sveltejs/kit": "^2.20.8",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@tailwindcss/typography": "^0.5.15",
     "@tailwindcss/vite": "^4.0.0",
-    "@tanstack/eslint-plugin-query": "^5.73.3",
-    "@tanstack/svelte-query": "^5.74.4",
-    "@tanstack/svelte-query-devtools": "^5.74.6",
+    "@tanstack/eslint-plugin-query": "^5.74.7",
+    "@tanstack/svelte-query": "^5.75.7",
+    "@tanstack/svelte-query-devtools": "^5.75.7",
     "@trivago/prettier-plugin-sort-imports": "^5.2.2",
     "@types/eslint": "^9.6.1",
+    "@types/qs": "^6.9.18",
     "autoprefixer": "^10.4.21",
     "eslint": "^9.25.1",
     "eslint-plugin-svelte": "^3.5.1",
@@ -35,28 +36,29 @@
     "prettier": "^3.5.3",
     "prettier-plugin-svelte": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
-    "svelte": "^4.2.19",
-    "svelte-check": "^3.8.6",
+    "svelte": "^5.28.2",
+    "svelte-check": "^4.1.7",
     "tailwind-merge": "^3.2.0",
     "tailwind-variants": "^1.0.0",
     "tailwindcss": "^4.0.0",
     "tslib": "^2.8.1",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.31.0",
-    "vite": "^5.4.18"
+    "typescript-eslint": "^8.32.0",
+    "vite": "^6.3.5"
   },
   "type": "module",
   "dependencies": {
-    "@stytch/vanilla-js": "^5.22.6",
-    "axios": "^1.8.4",
-    "bits-ui": "0.22.0",
+    "@stytch/vanilla-js": "^5.23.4",
+    "axios": "^1.9.0",
+    "bits-ui": "^1.4.7",
     "clsx": "^2.1.1",
     "dotenv": "^16.5.0",
-    "lucide-svelte": "^0.469.0",
-    "maplibre-gl": "^4.7.1",
+    "lucide-svelte": "^0.509.0",
+    "maplibre-gl": "^5.5.0",
     "oidc-client-ts": "^3.2.0",
-    "pmtiles": "^3.2.1",
-    "svelte-maplibre": "^0.9.14",
-    "zod": "^3.24.3"
+    "pmtiles": "^4.3.0",
+    "qs": "^6.14.0",
+    "svelte-maplibre": "^1.0.0",
+    "zod": "^3.24.4"
   }
 }
