@@ -1,15 +1,15 @@
 <script lang="ts">
-  import { Avatar as AvatarPrimitive } from "bits-ui";
   import { cn } from "$lib/utils.js";
+  import { Avatar as AvatarPrimitive } from "bits-ui";
 
-  type $$Props = AvatarPrimitive.FallbackProps;
+  type $Props = AvatarPrimitive.FallbackProps;
 
-  let className: $$Props["class"] = undefined;
+  let className: $Props["class"] = undefined;
   export { className as class };
 </script>
 
 <AvatarPrimitive.Fallback
-  class={cn("flex h-full w-full items-center justify-center rounded-full bg-muted", className)}
+  class={cn("bg-muted flex h-full w-full items-center justify-center rounded-full", className)}
   {...$$restProps}
 >
   <slot />
