@@ -1,12 +1,12 @@
 <script lang="ts">
-  import { Avatar as AvatarPrimitive } from "bits-ui";
   import { cn } from "$lib/utils.js";
+  import { Avatar as AvatarPrimitive } from "bits-ui";
 
-  type $$Props = AvatarPrimitive.ImageProps;
+  type $Props = AvatarPrimitive.ImageProps;
 
-  let className: $$Props["class"] = undefined;
-  export let src: $$Props["src"] = undefined;
-  export let alt: $$Props["alt"] = undefined;
+  let className: $Props["class"] = undefined;
+  export let src: $Props["src"] = undefined;
+  export let alt: $Props["alt"] = undefined;
   export { className as class };
 </script>
 
