@@ -1,11 +1,11 @@
 <script lang="ts">
-  import { Avatar as AvatarPrimitive } from "bits-ui";
   import { cn } from "$lib/utils.js";
+  import { Avatar as AvatarPrimitive } from "bits-ui";
 
-  type $$Props = AvatarPrimitive.Props;
+  type $Props = AvatarPrimitive.Props;
 
-  let className: $$Props["class"] = undefined;
-  export let delayMs: $$Props["delayMs"] = undefined;
+  let className: $Props["class"] = undefined;
+  export let delayMs: $Props["delayMs"] = undefined;
   export { className as class };
 </script>
 
