@@ -1,14 +1,15 @@
 <script lang="ts">
-  import { Button as ButtonPrimitive } from "bits-ui";
-  import { type Props, buttonVariants } from "./index.js";
   import { cn } from "$lib/utils.js";
+  import { Button as ButtonPrimitive } from "bits-ui";
 
-  type $$Props = Props;
+  import { type Props, buttonVariants } from "./index.js";
 
-  let className: $$Props["class"] = undefined;
-  export let variant: $$Props["variant"] = "default";
-  export let size: $$Props["size"] = "default";
-  export let builders: $$Props["builders"] = [];
+  type $Props = Props;
+
+  let className: $Props["class"] = undefined;
+  export let variant: $Props["variant"] = "default";
+  export let size: $Props["size"] = "default";
+  export let builders: $Props["builders"] = [];
   export { className as class };
 </script>
 
