@@ -1,10 +1,11 @@
 <script lang="ts">
   import type { HTMLAttributes } from "svelte/elements";
+
   import { cn } from "$lib/utils.js";
 
-  type $$Props = HTMLAttributes<HTMLDivElement>;
+  type $Props = HTMLAttributes<HTMLDivElement>;
 
-  let className: $$Props["class"] = undefined;
+  let className: $Props["class"] = undefined;
   export { className as class };
 </script>
 
