@@ -1,13 +1,14 @@
 <script lang="ts">
   import type { HTMLAttributes } from "svelte/elements";
+
   import { cn } from "$lib/utils.js";
 
-  type $$Props = HTMLAttributes<HTMLParagraphElement>;
+  type $Props = HTMLAttributes<HTMLParagraphElement>;
 
-  let className: $$Props["class"] = undefined;
+  let className: $Props["class"] = undefined;
   export { className as class };
 </script>
 
-<p class={cn("text-sm text-muted-foreground", className)} {...$$restProps}>
+<p class={cn("text-muted-foreground text-sm", className)} {...$$restProps}>
   <slot />
 </p>
