@@ -1,20 +1,22 @@
 <script lang="ts">
   import type { HTMLAttributes } from "svelte/elements";
-  import type { HeadingLevel } from "./index.js";
+
   import { cn } from "$lib/utils.js";
 
-  type $$Props = HTMLAttributes<HTMLHeadingElement> & {
+  import type { HeadingLevel } from "./index.js";
+
+  type $Props = HTMLAttributes<HTMLHeadingElement> & {
     tag?: HeadingLevel;
   };
 
-  let className: $$Props["class"] = undefined;
-  export let tag: $$Props["tag"] = "h3";
+  let className: $Props["class"] = undefined;
+  export let tag: $Props["tag"] = "h3";
   export { className as class };
 </script>
 
 <svelte:element
   this={tag}
-  class={cn("text-lg font-semibold leading-none tracking-tight", className)}
+  class={cn("text-lg leading-none font-semibold tracking-tight", className)}
   {...$$restProps}
 >
   <slot />
