@@ -1,15 +1,16 @@
 <script lang="ts">
   import type { HTMLAttributes } from "svelte/elements";
+
   import { cn } from "$lib/utils.js";
 
-  type $$Props = HTMLAttributes<HTMLDivElement>;
+  type $Props = HTMLAttributes<HTMLDivElement>;
 
-  let className: $$Props["class"] = undefined;
+  let className: $Props["class"] = undefined;
   export { className as class };
 </script>
 
 <div
-  class={cn("rounded-lg border bg-card text-card-foreground shadow-sm", className)}
+  class={cn("bg-card text-card-foreground rounded-lg border shadow-sm", className)}
   {...$$restProps}
 >
   <slot />
