@@ -1,15 +1,16 @@
 <script lang="ts">
-	import { Collapsible as CollapsiblePrimitive } from "bits-ui";
-	import { slide } from "svelte/transition";
+  import { slide } from "svelte/transition";
 
-	type $$Props = CollapsiblePrimitive.ContentProps;
+  import { Collapsible as CollapsiblePrimitive } from "bits-ui";
 
-	export let transition: $$Props["transition"] = slide;
-	export let transitionConfig: $$Props["transitionConfig"] = {
-		duration: 150,
-	};
+  type $Props = CollapsiblePrimitive.ContentProps;
+
+  export let transition: $Props["transition"] = slide;
+  export let transitionConfig: $Props["transitionConfig"] = {
+    duration: 150,
+  };
 </script>
 
 <CollapsiblePrimitive.Content {transition} {transitionConfig} {...$$restProps}>
-	<slot />
+  <slot />
 </CollapsiblePrimitive.Content>
