@@ -1,20 +1,20 @@
 <script lang="ts">
-  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
   import { cn } from "$lib/utils.js";
+  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
 
-  type $$Props = DropdownMenuPrimitive.ItemProps & {
+  type $Props = DropdownMenuPrimitive.ItemProps & {
     inset?: boolean;
   };
-  type $$Events = DropdownMenuPrimitive.ItemEvents;
+  type $Events = DropdownMenuPrimitive.ItemEvents;
 
-  let className: $$Props["class"] = undefined;
-  export let inset: $$Props["inset"] = undefined;
+  let className: $Props["class"] = undefined;
+  export let inset: $Props["inset"] = undefined;
   export { className as class };
 </script>
 
 <DropdownMenuPrimitive.Item
   class={cn(
-    "relative flex cursor-default select-none items-center rounded-sm px-2 py-1.5 text-sm outline-none data-[disabled]:pointer-events-none data-[highlighted]:bg-accent data-[highlighted]:text-accent-foreground data-[disabled]:opacity-50",
+    "data-[highlighted]:bg-accent data-[highlighted]:text-accent-foreground relative flex cursor-default items-center rounded-sm px-2 py-1.5 text-sm outline-none select-none data-[disabled]:pointer-events-none data-[disabled]:opacity-50",
     inset && "pl-8",
     className,
   )}
