@@ -1,13 +1,13 @@
 <script lang="ts">
-  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
   import { cn } from "$lib/utils.js";
+  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
 
-  type $$Props = DropdownMenuPrimitive.LabelProps & {
+  type $Props = DropdownMenuPrimitive.LabelProps & {
     inset?: boolean;
   };
 
-  let className: $$Props["class"] = undefined;
-  export let inset: $$Props["inset"] = undefined;
+  let className: $Props["class"] = undefined;
+  export let inset: $Props["inset"] = undefined;
   export { className as class };
 </script>
 
