@@ -1,9 +1,9 @@
 <script lang="ts">
   import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
 
-  type $$Props = DropdownMenuPrimitive.RadioGroupProps;
+  type $Props = DropdownMenuPrimitive.RadioGroupProps;
 
-  export let value: $$Props["value"] = undefined;
+  export let value: $Props["value"] = undefined;
 </script>
 
 <DropdownMenuPrimitive.RadioGroup {...$$restProps} bind:value>
