@@ -1,19 +1,19 @@
 <script lang="ts">
+  import { cn } from "$lib/utils.js";
   import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
   import Circle from "lucide-svelte/icons/circle";
-  import { cn } from "$lib/utils.js";
 
-  type $$Props = DropdownMenuPrimitive.RadioItemProps;
-  type $$Events = DropdownMenuPrimitive.RadioItemEvents;
+  type $Props = DropdownMenuPrimitive.RadioItemProps;
+  type $Events = DropdownMenuPrimitive.RadioItemEvents;
 
-  let className: $$Props["class"] = undefined;
-  export let value: $$Props["value"];
+  let className: $Props["class"] = undefined;
+  export let value: $Props["value"];
   export { className as class };
 </script>
 
 <DropdownMenuPrimitive.RadioItem
   class={cn(
-    "relative flex cursor-default select-none items-center rounded-sm py-1.5 pl-8 pr-2 text-sm outline-none data-[disabled]:pointer-events-none data-[highlighted]:bg-accent data-[highlighted]:text-accent-foreground data-[disabled]:opacity-50",
+    "data-[highlighted]:bg-accent data-[highlighted]:text-accent-foreground relative flex cursor-default items-center rounded-sm py-1.5 pr-2 pl-8 text-sm outline-none select-none data-[disabled]:pointer-events-none data-[disabled]:opacity-50",
     className,
   )}
   {value}
