@@ -1,14 +1,14 @@
 <script lang="ts">
-  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
   import { cn } from "$lib/utils.js";
+  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
 
-  type $$Props = DropdownMenuPrimitive.SeparatorProps;
+  type $Props = DropdownMenuPrimitive.SeparatorProps;
 
-  let className: $$Props["class"] = undefined;
+  let className: $Props["class"] = undefined;
   export { className as class };
 </script>
 
 <DropdownMenuPrimitive.Separator
-  class={cn("-mx-1 my-1 h-px bg-muted", className)}
+  class={cn("bg-muted -mx-1 my-1 h-px", className)}
   {...$$restProps}
 />
