@@ -1,13 +1,13 @@
 <script lang="ts">
-  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
   import { cn, flyAndScale } from "$lib/utils.js";
+  import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
 
-  type $$Props = DropdownMenuPrimitive.SubContentProps;
-  type $$Events = DropdownMenuPrimitive.SubContentEvents;
+  type $Props = DropdownMenuPrimitive.SubContentProps;
+  type $Events = DropdownMenuPrimitive.SubContentEvents;
 
-  let className: $$Props["class"] = undefined;
-  export let transition: $$Props["transition"] = flyAndScale;
-  export let transitionConfig: $$Props["transitionConfig"] = {
+  let className: $Props["class"] = undefined;
+  export let transition: $Props["transition"] = flyAndScale;
+  export let transitionConfig: $Props["transitionConfig"] = {
     x: -10,
     y: 0,
   };
@@ -18,7 +18,7 @@
   {transition}
   {transitionConfig}
   class={cn(
-    "z-50 min-w-[8rem] rounded-md border bg-popover p-1 text-popover-foreground shadow-lg focus:outline-none",
+    "bg-popover text-popover-foreground z-50 min-w-[8rem] rounded-md border p-1 shadow-lg focus:outline-none",
     className,
   )}
   {...$$restProps}
