@@ -1,21 +1,21 @@
 <script lang="ts">
+  import { cn } from "$lib/utils.js";
   import { DropdownMenu as DropdownMenuPrimitive } from "bits-ui";
   import ChevronRight from "lucide-svelte/icons/chevron-right";
-  import { cn } from "$lib/utils.js";
 
-  type $$Props = DropdownMenuPrimitive.SubTriggerProps & {
+  type $Props = DropdownMenuPrimitive.SubTriggerProps & {
     inset?: boolean;
   };
-  type $$Events = DropdownMenuPrimitive.SubTriggerEvents;
+  type $Events = DropdownMenuPrimitive.SubTriggerEvents;
 
-  let className: $$Props["class"] = undefined;
-  export let inset: $$Props["inset"] = undefined;
+  let className: $Props["class"] = undefined;
+  export let inset: $Props["inset"] = undefined;
   export { className as class };
 </script>
 
 <DropdownMenuPrimitive.SubTrigger
   class={cn(
-    "flex cursor-default select-none items-center rounded-sm px-2 py-1.5 text-sm outline-none data-[highlighted]:bg-accent data-[state=open]:bg-accent data-[highlighted]:text-accent-foreground data-[state=open]:text-accent-foreground",
+    "data-[highlighted]:bg-accent data-[state=open]:bg-accent data-[highlighted]:text-accent-foreground data-[state=open]:text-accent-foreground flex cursor-default items-center rounded-sm px-2 py-1.5 text-sm outline-none select-none",
     inset && "pl-8",
     className,
   )}
