@@ -1,11 +1,12 @@
 <script lang="ts">
   import type { HTMLAttributes } from "svelte/elements";
+
   import { cn } from "$lib/utils.js";
 
-  type $$Props = HTMLAttributes<HTMLDivElement>;
+  type $Props = HTMLAttributes<HTMLDivElement>;
 
-  let className: $$Props["class"] = undefined;
+  let className: $Props["class"] = undefined;
   export { className as class };
 </script>
 
-<div class={cn("animate-pulse rounded-md bg-muted", className)} {...$$restProps}></div>
+<div class={cn("bg-muted animate-pulse rounded-md", className)} {...$$restProps}></div>
