@@ -1,33 +1,29 @@
 <script lang="ts">
-  import maplibre, { type Map } from "maplibre-gl";
-  import { MapLibre } from "svelte-maplibre";
-  import { QueryClientProvider } from "@tanstack/svelte-query";
-  import "@fontsource-variable/inter";
-  import "../app.css";
-  import { SvelteQueryDevtools } from "@tanstack/svelte-query-devtools";
-  import MainCard from "$lib/components/MainCard.svelte";
-  import Markers from "$lib/components/Markers.svelte";
   import { setContext } from "svelte";
-  
-  import {
-    INITIAL_BEARING,
-    INITIAL_CENTER,
-    INITIAL_PITCH,
-    INITIAL_ZOOM
-  } from "$lib/constants";
-  
+
+  import { MapLibre } from "svelte-maplibre";
   import { writable } from "svelte/store";
-  import { Protocol } from "pmtiles";
-  import { page } from "$app/stores";
+
+  import { page } from "$app/state";
   import { queryClient } from "$lib/api";
   import LoginUserHeader from "$lib/components/LoginUserHeader.svelte";
-  
+  import MainCard from "$lib/components/MainCard.svelte";
+  import Markers from "$lib/components/Markers.svelte";
+  import { INITIAL_BEARING, INITIAL_CENTER, INITIAL_PITCH, INITIAL_ZOOM } from "$lib/constants";
+  import "@fontsource-variable/inter";
+  import { QueryClientProvider } from "@tanstack/svelte-query";
+  import { SvelteQueryDevtools } from "@tanstack/svelte-query-devtools";
+  import maplibre, { type Map } from "maplibre-gl";
+  import { Protocol } from "pmtiles";
+
+  import "../app.css";
+
   const protocol = new Protocol();
-  
+
   maplibre.addProtocol("pmtiles", protocol.tile);
-  
+
   const map = writable<Map>();
-  
+
   setContext("map", map);
 </script>
 
@@ -60,7 +56,7 @@
   <LoginUserHeader />
 
   {#if $map}
-    {#key $page.params.toiletId}
+    {#key page.params.toiletId}
       <MainCard>
         <slot />
       </MainCard>
